add abono to alquiler

+++ db/migrate/20120806011430_create_alquilers.rb
@@ -11,7 +11,7 @@ class CreateAlquilers < ActiveRecord::Migration
       t.integer :big_tables
       t.integer :tablecloth
       t.integer :big_tablecloth
-      t.float :total
+      t.integer :total
       t.timestamps
     end
   end
